Refactor (simplify) filter interfaces: Remove IXLAutoFilter. Rename IXLBaseAutoFilter to IXLAutoFilter. Remove IXLAutoFilter.Set method.
