+++ ClosedXML/Excel/AutoFilters/IXLAutoFilter.cs
@@ -1,16 +1,25 @@
 // Keep this file CodeMaid organised and cleaned
 using System;
+using System.Collections.Generic;
 
 namespace ClosedXML.Excel
 {
-    using System.Collections.Generic;
+    public enum XLFilterDynamicType { AboveAverage, BelowAverage }
+
+    public enum XLFilterType { Regular, Custom, TopBottom, Dynamic, DateTimeGrouping }
+
+    public enum XLTopBottomPart { Top, Bottom }
 
     public interface IXLAutoFilter
     {
+        Boolean Enabled { get; set; }
+        IXLRange Range { get; set; }
         Int32 SortColumn { get; set; }
         Boolean Sorted { get; set; }
         XLSortOrder SortOrder { get; set; }
         IEnumerable<IXLRangeRow> VisibleRows { get; }
+
+        IXLAutoFilter Clear();
 
         IXLFilterColumn Column(String column);
