+++ ClosedXML/Excel/Tables/XLTable.cs
@@ -176,7 +176,7 @@ namespace ClosedXML.Excel
             }
         }
 
-        public new IXLBaseAutoFilter SetAutoFilter()
+        public new IXLAutoFilter SetAutoFilter()
         {
             return AutoFilter;
         }
@@ -526,7 +526,7 @@ namespace ClosedXML.Excel
             return this;
         }
 
-        IXLBaseAutoFilter IXLTable.AutoFilter
+        IXLAutoFilter IXLTable.AutoFilter
         {
             get { return AutoFilter; }
         }
